pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sc-heimdall"
dynamic = ["version"]

authors = [
  { name="Remy Liu"},
  { name="Nicholas Ho"},
  { name="Shahul Alam", email="alam.shahul@gmail.com" },
]

dependencies = [  # TODO: check which are necessary
    "accelerate",
    "anndata>=0.11.0",
    "hydra-core",
    "numpy",
    "omegaconf",
    "packaging",
    "pandas",
    "wandb",
    "tqdm",
    "scanpy",
    "scikit-learn",
    "scipy",
    "torchmetrics",
    "transformers",
    "datasets",
    "einops",
    "mygene",
    "awkward",
]
description = "Heimdall: A Comprehensive Paradigm for Evaluating Single-Cell Representations within Foundational Models"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.hatch.build.targets.sdist]
include = [
  "Heimdall/*.py",
  "Heimdall/config/**",
]

[tool.hatch.build.targets.wheel]
include = [
  "Heimdall/*.py",
  "Heimdall/config/**",
]

[tool.hatch.version]
path = "Heimdall/__about__.py"

[tool.flake8]
max-line-length = 120
count = true
extend-ignore = [
    "E203",  # conflicting with black style slicing
    "D107",  # only need docstring at class level, no need at __init__
]
exclude = [
    ".nox",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.pyc",
    "*.egg-info",
    ".eggs",
]

[tool.black]
line-length = 120
preview = true

[tool.isort]
profile = "black"
line_length = 120
